requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.6.0
pydantic==2.5.2
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1 